pyproject.toml
[project]
name = "shopping-platform"
version = "0.1.0"
description = "E-commerce shopping platform built with Flask"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
